Use AJAX data source for puzzles table
