+++ hunts/templates/all_puzzles.html
@@ -11,7 +11,7 @@
     </button>
 </div>
 
-{% get_table puzzles request %}
+{% get_table hunt_pk request %}
 
 <div class="modal fade" id="addpuzzle" tabindex="-1" role="dialog" aria-labelledby="addtitle" aria-hidden="true">
     <div class="modal-dialog" role="document">
