+++ hunts/views.py
@@ -3,11 +3,12 @@ from django.contrib.auth.decorators import login_required
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.contrib import messages
 from django.db import IntegrityError
-from django.http import HttpResponseRedirect
+from django.http import HttpResponseRedirect, JsonResponse
 from django.shortcuts import render, get_object_or_404
 from django.utils.decorators import method_decorator
 from django.views import View
 from django.views.decorators.csrf import csrf_exempt
+from django.views.decorators.http import require_GET
 
 from url_normalize import url_normalize
 
@@ -16,6 +17,7 @@ from .models import Hunt
 from google_api_lib.google_api_client import GoogleApiClient
 from puzzles.forms import PuzzleForm
 from puzzles.models import Puzzle
+from puzzles.puzzle_tree import PuzzleTree
 from slack_lib.slack_client import SlackClient
 
 
@@ -43,6 +45,69 @@ def index(request):
 @login_required(login_url='/accounts/login/')
 def tools(request):
     return render(request, 'tools.html')
+
+
+def __table_status_class(puzzle):
+    if puzzle.status == Puzzle.PENDING:
+        return "table-warning"
+    elif puzzle.status == Puzzle.EXTRACTION:
+        return "table-danger"
+    elif puzzle.status == Puzzle.SOLVED:
+        return "table-success"
+    elif puzzle.status == Puzzle.STUCK:
+        return "table-danger"
+    else:
+        return ""
+
+
+def __get_puzzle_class(sorted_np_pairs):
+    puzzle_class = [__table_status_class(pair.node.puzzle) for pair in sorted_np_pairs]
+    most_recent_treegrid_id = {}
+    for i, pair in enumerate(sorted_np_pairs):
+        node = pair.node
+        parent = pair.parent
+        treegrid_id = i+1
+        most_recent_treegrid_id[node.puzzle.pk] = treegrid_id
+        puzzle_class[i] += " treegrid-%i" % treegrid_id
+        if parent and parent.puzzle != None:
+            # Most recently seen treegrid_id for parent
+            parent_treegrid_id = most_recent_treegrid_id.get(parent.puzzle.pk, 0)
+            if parent_treegrid_id > 0:
+                puzzle_class[i] += " treegrid-parent-%i" % parent_treegrid_id
+        # If meta is solved, collapse the subtree.
+        if len(node.children) > 0 and node.puzzle.is_solved():
+            puzzle_class[i] += " initial-collapsed"
+    return puzzle_class
+
+
+@require_GET
+@login_required(login_url='/accounts/login/')
+def puzzles(request, pk):
+    hunt = get_object_or_404(Hunt, pk=pk)
+    puzzle_objects = (hunt.puzzles.all().prefetch_related('metas')
+                                        .prefetch_related('active_users')
+                                        .prefetch_related('tags'))
+
+    sorted_np_pairs = PuzzleTree(puzzle_objects).get_sorted_node_parent_pairs()
+    sorted_puzzles = [pair.node.puzzle for pair in sorted_np_pairs]
+    puzzle_classes = __get_puzzle_class(sorted_np_pairs)
+
+    rows = zip(sorted_puzzles, puzzle_classes)
+
+    result = {
+        'data': [
+            [
+                puzzle.pk, puzzle.name, puzzle.url, puzzle.is_meta,
+                ['%s %s' % (user.first_name, user.last_name) for user in puzzle.active_users.all()],
+                puzzle.answer, puzzle.status, puzzle.sheet, puzzle.channel,
+                [[tag.name, tag.color] for tag in puzzle.tags.all()], '',
+                puzzle_class
+            ]
+            for puzzle, puzzle_class in rows
+        ]
+    }
+
+    return JsonResponse(result)
 
 
 class HuntView(LoginRequiredMixin, View):
